% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmistogram.R
\docType{data}
\name{SegmentShmistogram}
\alias{SegmentShmistogram}
\title{A class of methods for generating a shmistogram}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SegmentShmistogram
}
\description{
A class of methods for generating a shmistogram
}
\details{
"Segment" refers to the line segements used to represent
the multinomial component of the underlying density model
}
\keyword{datasets}
