% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmistogram.R
\name{shmistogram}
\alias{shmistogram}
\title{Plot a shmistogram}
\usage{
shmistogram(data, params = NULL, plotting_params = list(),
  return_data = FALSE)
}
\arguments{
\item{data}{A numeric vector. May contain NA}

\item{params}{List of parameters to pass to the shmistogram
constructor}

\item{plotting_params}{List of parameters passed to the plot call}

\item{return_data}{(boolean) If TRUE, return the shmistogram data
in addition to building the plot}
}
\description{
Plot a shmistogram
}
